@@ -1,51 +1,62 @@
 {
   "name": "personal-site",
-  "version": "3.0.1",
-  "homepage": "https://mldangelo.com/",
+  "version": "4.0.0",
+  "homepage": "https://victorolade.com/",
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/mldangelo/personal-site.git"
+    "url": "https://github.com/volade/personal-site.git"
+  },
+  "engines": {
+    "node": ">=16.x"
   },
   "scripts": {
+    "analyze": "npm run build && source-map-explorer build/static/js/*.chunk.js",
     "build": "rimraf ./build && react-scripts build",
+    "lint": "eslint \"**/*.js\"",
     "predeploy": "npm run build && react-snap",
-    "deploy": "gh-pages -d build",
     "start": "react-scripts start",
-    "lint": "eslint src",
-    "test": "react-scripts test --watchAll=false",
-    "analyze": "npm run build && source-map-explorer build/static/js/*.chunk.js"
+    "test": "npx jest"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.34",
-    "@fortawesome/free-brands-svg-icons": "^5.15.2",
-    "@fortawesome/free-regular-svg-icons": "^5.15.2",
-    "@fortawesome/react-fontawesome": "0.1.14",
-    "dayjs": "^1.10.4",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-brands-svg-icons": "^6.5.1",
+    "@fortawesome/free-regular-svg-icons": "^6.5.1",
+    "@fortawesome/react-fontawesome": "0.2.0",
+    "dayjs": "^1.11.10",
     "gh-pages": "^3.1.0",
-    "node-sass": "^4.14.1",
-    "prop-types": "^15.7.2",
-    "raw.macro": "^0.4.2",
-    "react": "^17.0.1",
-    "react-burger-menu": "^3.0.3",
-    "react-dom": "^17.0.1",
-    "react-ga": "^3.3.0",
-    "react-helmet-async": "^1.0.7",
+    "markdown-to-jsx": "^7.4.3",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-burger-menu": "^3.0.9",
+    "react-dom": "^18.2.0",
+    "react-ga": "^3.3.1",
+    "react-helmet-async": "^2.0.4",
+    "react-router-dom": "^6.22.3",
     "react-markdown": "^5.0.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.1",
+    "react-scripts": "^5.0.1",
+    "raw.macro": "^0.4.2",
     "react-snap": "^1.23.0",
     "rimraf": "^3.0.2"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.3",
-    "eslint": "^7.19.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "^7.22.0",
-    "source-map-explorer": "^2.5.2"
+    "@babel/eslint-parser": "^7.23.10",
+    "@babel/preset-env": "^7.24.0",
+    "@babel/preset-react": "^7.23.3",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.2",
+    "babel-jest": "^29.7.0",
+    "eslint": "^8.57.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "6.8.0",
+    "eslint-plugin-react": "^7.34.0",
+    "react-snap": "^1.23.0",
+    "rimraf": "^5.0.5",
+    "sass": "^1.71.1",
+    "source-map-explorer": "^2.5.3",
+    "@babel/plugin-proposal-private-property-in-object": "^7.14.5"
   },
   "browserslist": {
     "production": [
