@@ -65,7 +65,7 @@ class Skills extends Component {
       buttons.All = !Object.keys(prevState.buttons).some((key) => buttons[key]);
       return { buttons };
     });
-  }
+  };
 
   render() {
     return (
@@ -73,9 +73,6 @@ class Skills extends Component {
         <div className="link-to" id="skills" />
         <div className="title">
           <h3>Skills</h3>
-          <p>Note: I think these sections are silly, but everyone seems to have one.
-            Here is a *mostly* honest overview of my skills.
-          </p>
         </div>
         <div className="skill-button-container">
           {this.getButtons()}
