@@ -1,108 +1,69 @@
 const positions = [
   {
-    company: 'Arthena',
-    position: 'Co-founder, C.T.O.',
-    link: 'https://arthena.com',
-    daterange: 'January 2014 - Present',
+    company: 'Cart.com',
+    position: 'Director of Data Engineering',
+    link: 'https://cart.com',
+    daterange: 'January 2021 - Present',
     points: [
-      'Deployed quantitative strategies to predict the value of fine art in various pricing contexts.',
-      'Built production, on-line, end-to-end optimized machine learning pipelines with Pandas, Numpy, Scikit, Tensorflow, Dagster, Postgres, etc. on GCP with Kubernetes.',
-      'Designed micro-service architecture around: data collection, data integrity, feature engineering, research, strategy, backtesting, deployment, and reporting.',
-      'YC Alumn. Worked on everything. Built a 16-person team. Operated company for 6+ years.',
+      'Proficient in SQL for efficient querying within Snowflake and Big Query',
+      'Automated data preparation and sourcing using Python, fostering efficiency in workflows',
+      'Designed and implemented reporting dashboards utilizing Tableau',
+      'Engineered end-to-end ETL pipelines, incorporating modern technologies such as Snowflake, DBT, and Python',
+      'Leading a team of data engineers, providing mentorship and training in SQL, Python, JavaScript, and marketing analytics',
+      'Architected and implemented the data backend for a state-of-the-art analytics product, achieving full automation of ETL processes',
+      'Coordinated the seamless migration of the entire information infrastructure, ensuring uninterrupted service to clients during the transition',
+      'Extracted valuable insights from extensive and diverse datasets, transforming them into actionable data visualizations',
+      'Analyzed and interpreted complex data sets to identify trends and patterns, contributing to informed decision-making',
+      'Proactively identified and implemented data process improvements, enhancing overall operational efficiency',
+      'Recruited, trained, and managed both in-house and offshore teams, ensuring a collaborative and high-performance work environment',
     ],
   },
   {
-    company: 'Skeptical Investments',
-    position: 'Co-founder',
-    link: 'http://skepticalinvestments.biz',
-    daterange: 'April 2017 - Present',
+    company: 'Dumont Project',
+    position: 'Senior Manager, Business Intelligence',
+    link: 'cart.com',
+    daterange: 'April 2017 - July 2021',
     points: [
-      'Founded a micro-vc fund to throw token amounts of money at problems I find interesting.',
-      'Created the InstaSafe. Fast, automated investment docs for YC companies.',
+      'SQL writing, modeling, and querying relational databases on a daily basis',
+      'Writing python programs to automate data preparation and sourcing',
+      'Develop beautiful reporting dashboards leveraging DOMO, Tableau and other leading data visualization tools',
+      'Design, build, test and maintain data analytics solutions, making sure they meet business requirements and user needs',
+      'Build ETL pipelines, working with APIs and data warehouses',
+      'Mine, integrate and extract insights from large and disperse datasets and help translate them into clear and precise data visualizations',
+      'Help identify, analyze, and interpret trends or patterns in complex data sets',
+      'Locate and define new data process improvement opportunities',
+      'Recruit, train, supervise and manage in-house and offshore team',
     ],
   },
   {
-    company: 'Enveritas',
-    position: 'Consultant',
-    link: 'http://enveritas.org',
-    daterange: 'September 2016 - 2018',
+    company: 'ROI Media',
+    position: 'Ad Operations & Analytics',
+    link: '',
+    daterange: 'November 2016 - 2017',
     points: [
-      'Trained a model for the Brazilian Coffee Scenes Dataset with better than state of the art accuracy.',
-      'Collected training sets on the ground in Uganda. Built dashboards to visualize work of surveyors using Flask, React, and D3.',
-      'Performed supporting analysis to ensure data integrity using Pandas, t-SNE, SVM’s, and other techniques.',
+      'Prepared daily campaign optimization reports',
+      'Developed and maintained a state-of-the-art automated business intelligence platform from data extraction to visualization in QlikSense',
+      'Created data visualizations to answer most important business questions and explore trends',
+      'Develop and implement databases, data collection systems, data analytics and other strategies that optimize statistical efficiency and quality',
+      'Locate and define new process improvement opportunities',
+      'Prepare monthly financial reconciliation reports',
+      'Supervise support teams and source additional staff as needed',
     ],
   },
   {
-    company: 'Zenysis',
-    position: 'Consultant',
-    link: 'https://zenysis.com',
-    daterange: 'February 2016 - March 2016',
+    company: 'RAPP',
+    position: 'Ad Operations & Analytics',
+    link: 'https://rapp.com',
+    daterange: 'May 2015 - April 2016',
     points: [
-      'Worked in Addis Ababa for the Ethiopian Ministry of Health and built data visualization tools in React and Flask.',
-      'Ingested multiple databases with different alphabets, calendars, and without official spellings of geographic locations.',
-    ],
-  },
-  {
-    company: 'Matroid',
-    position: 'Co-founder',
-    link: 'https://matroid.com',
-    daterange: 'July 2015 - January 2016',
-    points: [
-      'Developed end to end machine learning pipeline to train visual classifiers from keywords using Node.JS, Express, Keystone, MongoDB, AWS, S3, Caffe, and other technologies.',
-      'Received Series A term sheets for 20M+ valuations.',
-    ],
-  },
-  {
-    company: 'Planet',
-    position: 'Missions Intern',
-    link: 'https://planet.com',
-    daterange: 'June 2014 - January 2015',
-    points: [
-      'Built models to improve image quality, signal to noise ratio, and dynamic range.',
-      'Performed statistical analysis of image quality in Matlab and Python.  Developed flight software in C++.',
-      'Organized first hackathon, prototyped hardware, and designed and built photography equipment for rocket launches.',
-    ],
-  },
-  {
-    company: 'Planetary Resources',
-    position: 'Avionics Intern',
-    link: 'http://planetaryresources.com',
-    daterange: 'January 2014 - May 2014',
-    points: [
-      'Developed simulations in Matlab for Attitude Determination and Control Subsystem.',
-      'Developed processes for in lab testing and characterization of various subsystems.',
-      'Assembled flight hardware in cleanroom.',
-    ],
-  },
-  {
-    company: 'Facebook',
-    position: 'Intern',
-    link: 'https://facebook.com',
-    daterange: 'June 2013 - September 2013',
-    points: [
-      'Developed software in python for automated testing of servers.',
-      'Performed statistical analysis with R, HIVE to assist in triage of malfunctioning servers.',
-      'Worked with vendors and ODM’s during triage to assist in risk mitigation.',
-    ],
-  },
-  {
-    company: 'SEDS-USA',
-    position: 'At Large Board Member',
-    link: 'http://seds.org',
-    daterange: 'October 2013 - October 2014',
-    points: [
-      'Elected to Board of Directors of the USA’s largest student space advocacy group based on 5+ years of work with SEDS.',
-      'Responsibilities included: organizational strategy, conference presentations, fundraising, special projects, promoting SEDS nationally and internationally, and photographing conferences.',
-    ],
-  },
-  {
-    company: 'UB Nanosatellite Program',
-    position: 'Co-founder, Program Manager',
-    link: 'https://ubnl.space/',
-    daterange: 'October 2010 – June 2012',
-    points: [
-      'Coauthored grant to build a multi-spectral imaging satellite as part of the AFRL University Nanosatellite Program.',
-      'Lead a team of 60 students through satellite development life cycle. Served as a technical expert, acquired intimate working knowledge of satellite subsystems.  Solicited funding through NASA, AFOSR, and several corporate sponsors.',
+      'Prepared weekly advertising stats and conducted market research',
+      'Coordinated with account managers on trafficking issues',
+      'Prepared monthly reconciliations',
+      'Resolved client issues in a timely manner',
+      'Resolved creative issues',
+      'Maintained / managed advertiser, vendor and partner communication',
+      'Supervised support teams and sourced additional staff as needed',
+      'Provided sales and budget reports as requested',
     ],
   },
 ];
