@@ -3,33 +3,38 @@
 
 const skills = [
   {
+    title: 'Big Query',
+    competency: 5,
+    category: ['Databases', 'Languages'],
+  },
+  {
+    title: 'Google Analytics',
+    competency: 5,
+    category: ['Tools'],
+  },
+  {
     title: 'Javascript',
     competency: 4,
     category: ['Web Development', 'Languages', 'Javascript'],
   },
   {
-    title: 'Node.JS',
-    competency: 3,
-    category: ['Web Development', 'Javascript'],
-  },
-  {
     title: 'React',
-    competency: 3,
+    competency: 2,
     category: ['Web Development', 'Javascript'],
   },
   {
-    title: 'Bash',
-    competency: 2,
+    title: 'Shell Scripting',
+    competency: 3,
     category: ['Tools', 'Languages'],
+  },
+  {
+    title: 'Snowflake',
+    competency: 5,
+    category: ['Tools'],
   },
   {
     title: 'Amazon Web Services',
     competency: 4,
-    category: ['Web Development', 'Tools'],
-  },
-  {
-    title: 'Heroku',
-    competency: 2,
     category: ['Web Development', 'Tools'],
   },
   {
@@ -38,18 +43,13 @@ const skills = [
     category: ['Web Development', 'Databases'],
   },
   {
-    title: 'ElasticSearch',
-    competency: 2,
-    category: ['Web Development', 'Databases'],
-  },
-  {
     title: 'PostgreSQL/SQLite3/SQL',
-    competency: 4,
+    competency: 5,
     category: ['Web Development', 'Databases', 'Languages'],
   },
   {
     title: 'Redis',
-    competency: 3,
+    competency: 4,
     category: ['Web Development', 'Databases'],
   },
   {
@@ -58,33 +58,28 @@ const skills = [
     category: ['Data Science'],
   },
   {
-    title: 'Express.JS',
-    competency: 2,
-    category: ['Web Development', 'Javascript'],
-  },
-  {
-    title: 'D3',
-    competency: 2,
-    category: ['Web Development', 'Javascript'],
-  },
-  {
     title: 'Flask',
-    competency: 2,
+    competency: 3,
     category: ['Web Development', 'Python'],
   },
   {
-    title: 'Git/Mercurial',
+    title: 'Django',
     competency: 3,
+    category: ['Web Development', 'Python'],
+  },
+  {
+    title: 'Git',
+    competency: 4,
     category: ['Tools'],
   },
   {
     title: 'Kubernetes',
-    competency: 2,
+    competency: 4,
     category: ['Tools', 'Data Engineering'],
   },
   {
     title: 'Google Cloud Compute',
-    competency: 2,
+    competency: 4,
     category: ['Tools', 'Web Development'],
   },
   {
@@ -93,24 +88,14 @@ const skills = [
     category: ['Data Science', 'Data Engineering', 'Python'],
   },
   {
-    title: 'Numba',
-    competency: 2,
-    category: ['Data Science', 'Data Engineering', 'Python'],
-  },
-  {
     title: 'Tensorflow + Keras',
-    competency: 3,
+    competency: 2,
     category: ['Data Science', 'Python'],
   },
   {
     title: 'Jupyter',
     competency: 3,
     category: ['Data Science', 'Python'],
-  },
-  {
-    title: 'Typescript',
-    competency: 2,
-    category: ['Web Development', 'Languages', 'Javascript'],
   },
   {
     title: 'HTML + SASS/SCSS/CSS',
@@ -121,26 +106,6 @@ const skills = [
     title: 'Python',
     competency: 5,
     category: ['Languages', 'Python'],
-  },
-  {
-    title: 'C++',
-    competency: 2,
-    category: ['Languages'],
-  },
-  {
-    title: 'Julia',
-    competency: 2,
-    category: ['Languages'],
-  },
-  {
-    title: 'MATLAB',
-    competency: 2,
-    category: ['Languages'],
-  },
-  {
-    title: 'R',
-    competency: 2,
-    category: ['Languages'],
   },
   {
     title: 'Data Visualization',
@@ -164,33 +129,13 @@ const skills = [
   },
   {
     title: 'Scikit-Learn',
-    competency: 4,
+    competency: 3,
     category: ['Data Engineering', 'Data Science', 'Python'],
   },
   {
-    title: 'Hadoop',
+    title: 'Cyber Security',
     competency: 2,
-    category: ['Data Engineering', 'Data Science'],
-  },
-  {
-    title: 'Spark',
-    competency: 2,
-    category: ['Data Engineering', 'Data Science'],
-  },
-  {
-    title: 'Dagster',
-    competency: 2,
-    category: ['Data Engineering', 'Python'],
-  },
-  {
-    title: 'Mypy',
-    competency: 3,
-    category: ['Python'],
-  },
-  {
-    title: 'Pylint',
-    competency: 4,
-    category: ['Data Engineering', 'Python'],
+    category: ['Web Development', 'Data Engineering', 'Data Science'],
   },
 ].map((skill) => ({ ...skill, category: skill.category.sort() }));
 
